+++ ios/RCTMGL/RCTMGLEventTypes.m
@@ -13,6 +13,8 @@
 NSString *const RCT_MAPBOX_EVENT_TAP = @"press";
 NSString *const RCT_MAPBOX_EVENT_LONGPRESS = @"longpress";
 
+NSString *const RCT_MAPBOX_USER_TRACKING_MODE_CHANGE = @"usertrackingmodechange";
+
 NSString *const RCT_MAPBOX_REGION_WILL_CHANGE_EVENT = @"regionwillchange";
 NSString *const RCT_MAPBOX_REGION_IS_CHANGING = @"regionischanging";
 NSString *const RCT_MAPBOX_REGION_DID_CHANGE = @"regiondidchange";